bug fix for encodings other than UTF-8
In Ruby a UTF-8 string "x" and a UTF-16 string "x" are not equal,
even when they consist of identical sequence of Unicode code points.
This cases problems on parser's token equality checks.  We need to
canonicalize the parser token in some way so here we take UTF-8.

A JSON-text must be a valid Unicode sequence so this must be OK.